pstl/omp: Breakup openmp backend into several files
This should make it easier to review the commit and to
maintain since the behavior is better categorized now.
